==> Machine Learning/10- ML Pipelines/01-withoutUsingPipelines.ipynb <==
#### This project is created to  focus on understanding and learning about pipelines and its working , rather than creating a best model.


#### This is the hassle way, without using transformers and all, in file 03.ipynb its shown , how to do it the easy and good way.This is for context.

import pandas as pd
import numpy as np

#### Importing Dataset

df= pd.read_csv('train.csv')

df.head()

df.info()

df.isnull().sum()

#### Removing the columns that we dont need now

df.drop(columns=['PassengerId','Name','Ticket','Cabin'],inplace=True)

X=df.drop(columns='Survived')
y=df['Survived']

#### Train Test split

from sklearn.model_selection import train_test_split
X_train,X_test,y_train,y_test= train_test_split(X,y,test_size=0.2,random_state=42)

X_train.head()

#### Applying Imputation to handle missing values. we need to apply imputation to Age column and embarked columns,since those column contain missiing values

# Applying imputation
# Will fill all the missing values 

from sklearn.impute import SimpleImputer

si_age = SimpleImputer() #### age-> with mean
si_embarked = SimpleImputer(strategy='most_frequent')  #### embarked-> witht the most frequently occuring value


#### Impute and Trasnfrom age and embark columns
X_train_age = si_age.fit_transform(X_train[['Age']])
X_train_embarked = si_embarked.fit_transform(X_train[['Embarked']])

X_test_age = si_age.transform(X_test[['Age']])
X_test_embarked = si_embarked.transform(X_test[['Embarked']])

#### Apply OneHotEncoding nominal- categorical column. In this case,we apply OneHotEncoding to  Sex and Embarked Column

from sklearn.preprocessing import OneHotEncoder

####Encode the Sex and Embarked column.
ohe_sex = OneHotEncoder(sparse_output=False,handle_unknown='ignore')
ohe_embarked = OneHotEncoder(sparse_output=False,handle_unknown='ignore')

####fit and transfrom the Sex and Embarked column. 
X_train_sex = ohe_sex.fit_transform(X_train[['Sex']])
X_train_embarked = ohe_embarked.fit_transform(X_train_embarked)

X_test_sex = ohe_sex.transform(X_test[['Sex']])
X_test_embarked = ohe_embarked.transform(X_test_embarked)

X_train_sex

X_train_embarked

#### so now we have three new arrays
#### for age, sex and embarked
#### X_train_age,X_train_sex,X_train_embarked

#### since we have new arrays for the age,sex and embarked column, we will remove the age, sex, embarked column, that we had and replace those with new ie X_train_age,X_train_sex,X_train_embarked

#### Drop the columns in train  and testing set

X_train_rem = X_train.drop(columns=['Sex','Age','Embarked']) 

X_test_rem = X_test.drop(columns=['Sex','Age','Embarked'])

#### Concatenate the training and testing dataset with new columns

X_train_transformed = np.concatenate((X_train_rem,X_train_age,X_train_sex,X_train_embarked),axis=1)
X_test_transformed = np.concatenate((X_test_rem,X_test_age,X_test_sex,X_test_embarked),axis=1)

X_train_transformed

X_train_transformed.shape

X_test_transformed.shape

#### Using DecisionTreeClassifier to train the model

from sklearn.tree import DecisionTreeClassifier
clf=DecisionTreeClassifier()
clf.fit(X_train_transformed,y_train)

#### Predicting with testing dataset

y_pred= clf.predict(X_test_transformed)

#### Checking the accuracy

from sklearn.metrics import accuracy_score
accuracy=accuracy_score(y_test,y_pred)

print(accuracy)

#### Exporting the model for future use case.

import pickle

pickle.dump(ohe_sex,open('models/ohe_sex.pkl','wb'))
pickle.dump(ohe_embarked,open('models/ohe_embarked.pkl','wb'))
pickle.dump(clf,open('models/clf.pkl','wb'))
