==> Machine Learning/10- ML Pipelines/03-usingPipelines.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.compose import ColumnTransformer
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import MinMaxScaler
from sklearn.pipeline import Pipeline,make_pipeline
from sklearn.feature_selection import SelectKBest,chi2
from sklearn.tree import DecisionTreeClassifier

df=pd.read_csv('train.csv')

df.head()

####Plan
1.Impute missing data in age and embark column using 1 column transformer
2.we will send that data to second column transformer, that will one hot encoded sex and emabark column
3. then the ouput from that , we will scale ,
4.then we will do feature selection, selecting best 5 column //not necessary, just for knowning how to do it
5.train model using decision tree

#### Drop the unncessary Columns

df.drop(columns=['PassengerId','Name','Ticket','Cabin'],inplace=True)

#### Train-test-split

X=df.drop(columns='Survived')
y=df['Survived']

X_train,X_test,y_train,y_test= train_test_split(X,y,test_size=0.2,random_state= 42)

X_train.head()

y_train

#### Imputation using Column Transformer to fill the missing values

##we will fill ,
##age->mean value
##emabrk->most frequent value

##Point to be noted, we need to pass the column index instead of name beacuse , the output that
##comes out is a numpy array, aarray doesnt have column names.
###if we pass column names,  it might  cause error
#### best to use index values
trf1=ColumnTransformer([
    ('impute_age',SimpleImputer(),[2]), #### passing the age column, index value, ie 2
    ('impute_embark',SimpleImputer(strategy='most_frequent'),[6]),#### passing the embark column, index value, ie 6
],remainder='passthrough')

####Encoding both columns at once
trf2=ColumnTransformer([
    ('ohe_sex_embarked',OneHotEncoder(sparse_output=False,handle_unknown='ignore'),[1,6]), #### passing the age and sex column, index value, ie 2
    #### passing the embark column, index value, ie [1,6]
],remainder='passthrough')

#### we are not doing dropfirst because we are using DecisionTree

#### Scaling

trf3=ColumnTransformer([
    ('scale',MinMaxScaler(),slice(0,10)) ####applied for all, would skip categorical col if wanted
    
])
#### we are using MinMaxscaler instead of standardScaler because
####we are doing feature selection
####we are doing only for this case

#### Feature Selection

#### for this project, its better to remove this step,  keeping it just to learn.
trf4=SelectKBest(score_func=chi2,k=8)

#### Train the model


trf5=DecisionTreeClassifier()

#### Create pipeline

#### better to use
pipe = Pipeline([
    ('trf1',trf1),
    ('trf2',trf2),
    ('trf3',trf3),
    ('trf4',trf4),
    ('trf5',trf5)
])

## Pipeline Vs make_pipeline

#### Pipeline requires naming of steps, make_pipeline does not.

#### (Same applies to ColumnTransformer vs make_column_transformer)

# Alternate Syntax 
#pipe = make_pipeline(trf1,trf2,trf3,trf4,trf5)

pipe.fit(X_train,y_train)

## Exploring

#### Pipe object has been traain now

pipe.named_steps  ## gives all steps that this particular pipeline follows

pipe.named_steps['trf1'].transformers_

pipe.named_steps['trf1'].transformers_[0]

pipe.named_steps['trf1'].transformers_[0][1]

pipe.named_steps['trf1'].transformers_[0][1].statistics_

pipe.named_steps['trf2'].transformers_

pipe.named_steps['trf1'].transformers_[1][1]

pipe.named_steps['trf1'].transformers_[1][1].statistics_

#### Predicting

y_pred= pipe.predict(X_test)

y_pred

from sklearn.metrics import accuracy_score
accuracy_score(y_test,y_pred)


## Cross Validation using Pipeline


# cross validation using cross_val_score
from sklearn.model_selection import cross_val_score
cross_val_score(pipe, X_train, y_train, cv=5, scoring='accuracy').mean()

#### GridSearch using Pipeline

params = {
    'trf5__max_depth':[1,2,3,4,5,None]
}

from sklearn.model_selection import GridSearchCV
grid = GridSearchCV(pipe, params, cv=5, scoring='accuracy')
grid.fit(X_train, y_train)

grid.best_score_

grid.best_params_

#### Exporting Pipeline

# export 
import pickle
pickle.dump(pipe,open('pipe.pkl','wb'))
